Introduce new 'multi' testnet.

+++ src/Gulden/Common/diff_delta.cpp
@@ -170,7 +170,7 @@ unsigned int GetNextWorkRequired_DELTA (const INDEX_TYPE pindexLast, const BLOCK
     }
 
     // -- Calculate time interval for middle window
-    if (INDEX_HEIGHT(pindexLast) - nFirstDeltaBlock <= nMiddleFrame)
+    if (INDEX_HEIGHT(pindexLast) - (int)nFirstDeltaBlock <= (int)nMiddleFrame)
     {
         nMiddleWeight = nMiddleTimespan = 0;
     }
@@ -196,7 +196,7 @@ unsigned int GetNextWorkRequired_DELTA (const INDEX_TYPE pindexLast, const BLOCK
 
     // -- Calculate timespan for long window
     // NB! No need to worry about single negative block times as it has no significant influence over this many blocks.
-    if (INDEX_HEIGHT(pindexLast) - nFirstDeltaBlock <= nLongFrame)
+    if ((int)INDEX_HEIGHT(pindexLast) - (int)nFirstDeltaBlock <= (int)nLongFrame)
     {
         nLongWeight = nLongTimespan = 0;
     }
